AiPerfTestAnalysis/requirements.txt
flask~=3.1.0
dotenv~=0.9.9
google-genai
pandas~=2.2.3
python-dotenv~=1.1.0
requests~=2.32.3